Merged in the changes from the alias branch into the compiler directory,
Estimated hours taken: 14 (mostly due to juggling revision tags, attempting
	to diagnose a CVS bug, and dealing with the repository's fascist
	attitude to copyright messages if you're a remote user)

Merged in the changes from the alias branch into the compiler directory,
up to the tag alias_snapshot_19990114150601.
Merged in the changes between the tag trunk_snapshot_19981111185105 and the
tag trunk_snapshot_19981123164219 for the rest of the system.

Note that while trunk_snapshot_19981123164219 is the trunk version
corresponding to the last commit on the alias branch before
alias_snapshot_19990114150601, the compiler won't bootcheck without a newer
version of the rest of the system.  I would have incorporated that change
into this commit, except that a bug in CVS is preventing the update.  I'm
hoping that by committing this change and then trying the update again, I
can work around the bug.