Brought the HAL branch of the compiler up-to-date with the alias branch.
Estimated hours taken: 6

Brought the HAL branch of the compiler up-to-date with the alias branch.

In the compiler directory, this involved merging in the changes between
the tags alias_snapshot_19990114150601 and alias_snapshot_19990520160750
(in particular, this included the merging of changes from the main
branch up to 3rd May 1999).  For all other directories, this meant
merging in the changes between the tags trunk_snapshot_19990120180856
and trunk_snapshot_19990520160750.

compiler/.hal_alias_sync_tag:
	New file recording the version of the alias branch with which we are
	up-to-date (this file existed already, but had failed to be
	committed to the repository).