Fix a problem on IRIX caused by my previous attempt to
Estimated hours taken: 2

Fix a problem on IRIX caused by my previous attempt to
fix a problem with <inttypes.h> on Solaris.  The problem
was that IRIX has <inttypes.h>, but its <inttypes.h>
does not define the types that we need.  The fix is to
get configure to explicitly check for the types that we
need, rather than just checking for the header file.

runtime/mercury_conf.h.in:
	Add new macros HAVE_SYS_TYPES, HAVE_INTPTR_T,
	and HAVE_INT_LEASTN_T.

configure.in:
	Add code to check for <sys/types.h>,
	{u,}intptr_t, and {u,}int_least{8,16,32}_t,
	and define HAVE_SYS_TYPES, HAVE_INTPTR_T,
	and HAVE_INT_LEASTN_T accordingly.

runtime/mercury_types.h:
	Define {u,}intptr_t and {u,}int_least{8,16,32}_t
	iff they aren't defined by the system header files,
	instead of assuming that they don't need to be
	defined if we have <stdint.h> or <inttypes.h>.