New versions of TeXinfo are more picky about the correct use of `@xref' and
Estimated hours taken: 3 (including time to learn enough about TeXinfo to
figure out what was going on)

New versions of TeXinfo are more picky about the correct use of `@xref' and
friends.  In particular, they insist that an `@xref' be followed immediately
by a comma or a full stop.  There were a few instances in the user manual
where this was not the case.

doc/user_guide.texi:
	Changed a few incorrect uses of `@xref' to use `@ref' instead, to
	keep texinfo happy.