Fix a bug that caused a variable not found error when compiling
Estimated hours taken: 1

Fix a bug that caused a variable not found error when compiling
library/term_io.m with intermodule optimization in my local version
of the Mercury compiler (which has various changes for existential types).
The problem was that when liveness.m calls requantify_proc,
quantification.m discovers that a variable is now local to a subgoal,
but leaves the instmap delta on the containing goal unchanged, still
containing a (now bogus) entry for that variable.
I think the same problem could occur without my existential types changes,
but constructing a test case to exercise this bug is difficult.

compiler/quantification.m:
	Ensure that the instmap delta is restricted to only contain
	entries for variables that occur in the goal's nonlocals.