This fixes two bugs in mmake's handling of split library files.
Estimated hours taken: 2

This fixes two bugs in mmake's handling of split library files.  The first
was that the `clean' target(s) did not remove the `.dir' directories for any
modules except the top-level one(s).  The second was that the split files
generated did not depend on the relevant interface and optimisation files,
so mmake sometimes tried to generate the split files with some necessary
files either missing or out-of-date.

compiler/modules.m:
	Changed an entry in the `clean' target generated in `.dep' files to
	remove `$(<module>.dirs)' rather than just `<module>.dir'.
	Added a dependency between the `.o' files generated by the split
	library targets and the relevant `.int', `.opt' and `.trans_opt'
	files.  As part of this, introduced a new predicate
	`module_name_to_split_c_file_pattern', which is like
	`module_name_to_split_c_file_name' but generates a wildcard pattern
	rather than a specific file name.