The code we were creating for the comparison procedure of an enumerated type
Estimated hours taken: 6

The code we were creating for the comparison procedure of an enumerated type
called mercury__compare_3_3, which called the comparison procedure of the
enumerated type (whose address was in the compare slot of the type_ctor_info).
This was an infinite loop. I think it only used to work because higher-order
specialization specialized the call to mercury__compare_3_3, but specialization
is turned off for debugging.

compiler/unify_proc.m:
	Fix the generation of the compare procedures for enumerated types;
	instead of calling mercury__compare_3_3, we cast the args to ints
	and call builtin_compare_int.

	At the same time, make the code that creates the special procedure
	bodies more robust by explicitly checking for invalid kinds of types.

compiler/make_hlds.m:
	Avoid a compiler abort from the newly strict sanity checks
	by not trying to create bodies for special preds for uu types;
	let make_hlds get to the code that reports an error instead.