Fix a problem where the compiler was leaking memory
Estimated hours taken: 16

Fix a problem where the compiler was leaking memory
when compiling more than one file a time.
Note that this fix solves that specific bug,
but does not address the problem in general.
(A more general fix would be for the garbage collector
to scan only the parts of the Mercury stacks which
are currently live, but this is trickier...)

compiler/mercury_compile.m:
	If we're processing more than one module, then invoke garbage
	collection explicitly after processing each one.

library/gc.m:
	New module.  Currently this only contains one routine,
	`garbage_collect', which forces a garbage collection.
	This routine clears out the Mercury stacks and registers
	before invoking GC_gcollect().

	Eventually we might want to include a more extensive interface,
	including e.g. features to enable and disable collection, etc.
	(see boehm_gc/gc.h for a list of useful features).

runtime/mercury_regs.h:
	Add macro MR_clear_regs_for_GC(), for use by gc__garbage_collect.

runtime/mercury_memory_zones.h:
	Add macro MR_clear_zone_for_GC(), for use by gc__garbage_collect.

NEWS:
	Mention the new `gc' library module.