Compile the files in the library directory with --trace minimum by default,
Estimated hours taken: 1

Compile the files in the library directory with --trace minimum by default,
which has no effect in non-debugging grades and causes the library to be
shallow traced, not deep traced, in debugging grades.

This is probably what most users want, and it makes it much easier to
maintain the expected output of the debugging test cases in debugging grades.

library/Mmakefile:
	Add --trace minimum to MCG.

tests/*/*:
	Update the test cases both for this change and for my previous change,
	the addition of line numbers.