Fix a bug where mdb was not printing line numbers for exception events.
Estimated hours taken: 4

Fix a bug where mdb was not printing line numbers for exception events.
The bug was that MR_trace_throw() was making a modified copy of the
return label stack layout, and passing the address of that copy
to MR_trace(), but in many places MR_trace() was depending on the
address passed to it being in the module layout table.

runtime/mercury_stack_layout.h:
	Define a new type MR_Exception_Layout, containing both
	the modified copy and the address of the original return label
	layout.

library/exception.m:
	Use the new type.  Store the modified copy of the layout in
	the first field.  Store the address of the original return label
	layout in the second field.  Pass the address of the first field
	to MR_trace().

trace/mercury_trace.c:
	Change MR_trace_real() and MR_trace_event() so that after
	extracting the port and path from the layout, they check for
	MR_PORT_EXCEPTION, and if so, (down)cast the layout pointer
	to `MR_Exception_Layout *', retrieve the return_label_layout
	from that struct, and use that for the new layout pointer.