Fix a bug in the term browser: the term size limit was by default
Estimated hours taken: 0.2

Fix a bug in the term browser: the term size limit was by default
being set to the default depth value, 10.  The depth value was
hard-coded to 3.  Since 3 is a more sensible depth limit (atleast
for the one-line printing that the internal debugger does), the default
has been changed to that.

browser/parse.m:
	Set the intended default depth to 3.

browser/browse.m:
	Set the browser defaults in a more readable fashion.
	Don't hard-code the default depth limit in the help message.