Avoid generating error messages for imported Aditi predicates
Estimated hours taken: 1.5

Avoid generating error messages for imported Aditi predicates
which are not used. The check for abstract types in the arguments
of Aditi relations must be performed for imported predicates which
are used because discriminated union types are written as abstract
types in `.int2' files. A type for which the body is available in
the module defining a predicate may be seen as an abstract type
in an importing module.

compiler/magic.m:
	Only process imported Aditi procedures which are used.

tests/valid/Mmakefile:
tests/valid/aditi_error_bug.m:
tests/valid/aditi_error_bug2.m:
tests/valid/aditi_error_bug3.m:
	Test case.