Avoid the cast to long before printing integers.
Estimated hours taken: 1

Avoid the cast to long before printing integers.  This cast is
problematic when printing negative integers with the format string "%x".
For instance on the alpha a long is 64 bits and a MR_integer is 32 bits,
so printing -31 is printed as ffffffffffffffe1 instead of ffffffe1.

configure.in:
    Determine the length modifier to use for MR_Integer.

runtime/mercury_conf.h.in:
    Record the length modifer using the #define
    MR_INTEGER_LENGTH_MODIFIER.

library/string.m:
    Add a new function int_length_modifier which returns the length
    modifier to use for a MR_Integer.
    Add an extra argument to make_format which is the length modifier to
    use when constructing the format string.