Alter the `dd' command in mdb so that it does not assume that the answer
Estimated hours taken: 2.5

Alter the `dd' command in mdb so that it does not assume that the answer
to the first question asked is `no'.  The question is somewhat redundant,
since the only sensible answer is `no', but I find that it gives useful
feedback about the symptom that is being diagnosed, and it makes clearer
why the following questions are asked.

browser/declarative_analyser.m:
	Start analysis with a list of suspects which contains only the
	topmost node of the debugging tree.

browser/declarative_debugger.m:
	If the answer to the first question is something other than `no',
	then end the diagnosis and report that no bugs were found.

tests/debugger/declarative/*.{inp,exp,exp2}:
	Update test cases to handle the extra question.