When exposing a Mercury predicate to C, you need to make sure there is
Estimated hours taken: 4

When exposing a Mercury predicate to C, you need to make sure there is
an engine for it to run in before calling it.

For single-threaded programs, we simply make sure that the engine is run
soon after main().  But for multi-threaded programs, code written in C
could create new threads and call Mercury using one of these threads.
Of course the same code could be called from a thread where there is
already a Mercury engine.

This change fixes this problem by creating a Mercury engine if and only
if it is required.

compiler/export.m:
	Run init_engine when calling into Mercury.  It is important to
	run this after we save the C registers, but before we try to use
	the Mercury registers.

runtime/mercury_thread.c:
runtime/mercury_thread.h:
	If an existing Mercury engine is already in the current thread,
	don't create a new one.