Fix a bug which caused a code generator abort for some procedures
Estimated hours taken: 1

Fix a bug which caused a code generator abort for some procedures
containing complicated in-in unifications. The problem was that
the type-infos added by simplify.m for the call to unify/2 were
not being added to the nonlocals set of the call. saved_vars.m
then removed the type-info variables because they were constants
local to the conjunction which were not used by any goal in the
conjunction.

compiler/simplify.m:
	Make sure type-info variables constructed for a call
	to unify/2 are included in the non-locals of the call.

tests/valid/Mmakefile:
tests/valid/unify_typeinfo_bug.m:
	Test case.