Fix a bug in the handling of typeclass method implementations which
Estimated hours taken: 0.25

Fix a bug in the handling of typeclass method implementations which
call field access functions. The code to build the clauses was putting
the wrong arity in the functor field of the unification created
to represent the function call. The type-checker does not look at
the arity for calls to ordinary user-defined functions.

compiler/make_hlds.m:
	Call construct_pred_or_func_call from produce_instance_method_clauses,
	rather than incorrectly duplicating the code.

tests/hard_coded/typeclasses/record_syntax.{m,exp}:
	Add a test for this bug.