Synchronise the HAL branch with the main branch.
Estimated hours taken: 50

Synchronise the HAL branch with the main branch.

These were mostly boring changes, a lot of which were related to using the
new RTTI representations.

The one real addition is that we now spit out a dummy ptag definition for
the `var' tag (in the appropriate grade) so that looking up a functor
descriptor from its tag now works in the reserve_tag grade.