The io__stream is used as a key in two maps in io.m.
Estimated hours taken: 1.5

The io__stream is used as a key in two maps in io.m.
This is a bad idea, io__stream is defined as a c_pointer, which should
not necessarily be comparable (there is no guarantee that it is even
constant -- a garbage collector might move it).

In the IL backend, this causes more trouble, as the .NET GC is a
compacting generational garbage collector, so in practice the
io__stream gets moved pretty soon after being allocated.

library/io.m:
	Add a new type -- the io__stream_id, which is a unique
	integer identifier for an IO stream, which can be extracted from
	the stream itself.
	Use io__stream_id as a key instead of io__stream.
	Implement the stream_id as simply the pointer itself in the
	boehm_gc C backend.
	(For the IL backend and the accurate GC backend the stream_id
	will be a unique integer stored in each stream).