Fix a problem with undefined symbols in the hl*prof* grades.
Estimated hours taken: 1

Fix a problem with undefined symbols in the hl*prof* grades.
This was precipitated by my recent change to the handling
of `:- external' for the MLDS back-end.  As a result of that
change, the compiler now reports link errors in hl*prof* grades
if there were any symbols declared as `:- external' and not
defined.

runtime/mercury_tabling.c:
	Add stub definitions of table_suspend and table_resume
	for the --high-level-code grades; currently these
	just call error/1.

library/private_builtin.m:
	Delete the `:- external' declaration for the builtin
	procedures `builtin_int_lt' and `builtin_int_gt'.
	That was needed for boostrapping a few years ago
	but is no longer necessary.

library/private_builtin.m:
	Provide a definition of free_heap, to avoid link errors
	in hl*prof* grades.