Add support for :- pragma foreign_proc -- this is a synonym for
Estimated hours taken: 1.0

Add support for :- pragma foreign_proc -- this is a synonym for
:- pragma foreign_code with arity > 2.

For the moment foreign_code with arity > 2 is still accepted, because
otherwise this will cause a lot of bootstrapping problems.

When the change is fully bootstrapped, we can remove support for :-
pragma foreign_code/3,4,5,etc altogether.

compiler/prog_io_pragma.m:
	Recognize the new pragma.