Add a whole bunch of entries to the .nocopyright files on the HAL branch.
Estimated hours taken: 1

Add a whole bunch of entries to the .nocopyright files on the HAL branch. This
is to avoid having to keep copyright messages up to date in a whole bunch
of files that haven't really changed. This branch will die soon anyway (when
the important changes are merged back on to the main branch), so there is no
point spending the time maintaining copyright messages.