Merge changes to the main branch from the last month or so onto the HAL
Estimated hours taken: 40

Merge changes to the main branch from the last month or so onto the HAL
branch.

Some of the support for the .se and .init grades has also been removed in
this merge. These grades are no longer needed and keeping them working would
have required quite a bit of extra work in this merge.

Note that I have added superfluous entries to the .nocopyright files in a lot
of directories to save time --- this branch will die soon anyway, so there's
no point spending the time to keep the copyright messages up to date.