Fix a bug: if `--trans-intermod-opt' was included in MCFLAGS,
Estimated hours taken: 2.5

Fix a bug: if `--trans-intermod-opt' was included in MCFLAGS,
then mmake would pass `--make-opt-int --trans-intermod-opt' to
mmc when creating the `.opt' files, and this would cause the
compiler to read the `.trans_opt' files when creating the `.opt'
files.  That's a bug, since the `.opt' files are not supposed to
depend on the `.trans_opt' files.

compiler/mercury_compile.m:
	Make sure that we don't read in the `.trans_opt' files
	when creating the `.opt' file.