Added an option `--optimize-tailcalls' which makes optimization of
Estimated hours taken: 1.

Added an option `--optimize-tailcalls' which makes optimization of
tailcalls in the MLDS backend optional.  This is turned off in
optimization levels -1 and 0 and when the Java grade is being used; for
other optimization levels it is on.

compiler/handle_options.m:
compiler/mercury_compile.m:
compiler/options.m:
doc/user_guide.texi:
	Added new option `--optimize-tailcalls'.