Fix a couple of bugs in the IL back-end that caused it to generate incorrect
Estimated hours taken: 2
Branches: main

Fix a couple of bugs in the IL back-end that caused it to generate incorrect
refences to constructor names when --high-level-data was enabled.
One was that it was using the base class type rather than the
derived type for the appropriate constructor.
The other was that it was using "." rather than "/" as the class qualifier
when referencing nested classes.
(This change is only a partial fix for the latter problem.)

compiler/ilds.m:
	Change the structured_type so that nested classes are stored
	a separate field from the namespace-qualified top-level class.

	Rename "append_class_name" as "append_toplevel_class_name",
	and add a new routine "append_nested_class_name".

compiler/ilasm.m:
	Output "/" rather than "." as the class qualifier for nested classes.

compiler/mlds_to_csharp.m:
compiler/mlds_to_mcpp.m:
	Minor changes to handle the new definition of structured_type.

compiler/mlds_to_il.m:
	Handle the new definition of structured_type.
	Use the MaybeCtorName field of new_object instructions
	to determine the type to allocate.