The compiler usually considers a file extension to include the `.',
Estimated hours taken: 0.25
Branches: main

The compiler usually considers a file extension to include the `.',
so pass `.o' rather than `o' as the value of `--object-file-extension'.
This change avoids the need to handle `--executable-file-extension' (which
can be empty) as a special case.

scripts/mmc.in:
	Pass `.o' as the value of `--object-file-extension', rather than `o'.

compiler/mercury_compile.m:
	Remove code to prepend a `.' to the value of `--object-file-extension'
	before passing it to `module_name_to_file_name'.

compiler/options.m:
	Update the default values for the various `--*-extension' options
	to include the `.'.