This is the first part of a change to make c2init and ml take
Estimated hours taken: 0.25

This is the first part of a change to make c2init and ml take
the same set of options (the c2init options will be added as
an "Initialization options" section of the ml options). When
this change is complete, users will only need to add `--trace'
to MLFLAGS, not C2INITFLAGS as well. c2init is an implementation
detail that users shouldn't need to know about.

scripts/c2init.in:
	Remove the `-l' short form of the `--library' option,
	since it conflicts with the use of `-l' as a linker option.

	Add `--no-main' as a synonym for `--library', which will
	cause less confusion with the `--make-shared-lib' ml
	option. `--library' will still be accepted for backwards
	compatability because the C interface samples refer to it.

	Add `--init-c-file' as a synonym for `-o' and `--output'.
	The latter two will be removed after mmake and mmc are
	changed to use `--init-c-file' because they don't make
	sense as ml options.