When compiling a module which contains a nested sub-module you must
Estimated hours taken: 1
Branches: main

When compiling a module which contains a nested sub-module you must
build the .int0 file before attempting to build the .int file.  This is
because while building the .int file you will need the .int0 file to
build the .int file of the nested sub-module.

compiler/modules.m:
    Add the dependency <module>.date : <module>.date0