Altered mlds_to_c.m so that it will print multiple return types when
Estimated hours taken: 1.0
Branches: main

Altered mlds_to_c.m so that it will print multiple return types when
performing an mlds dump. This case will not occur in the C back-end, but it
is useful to show this information when working on other back-ends such as
Java.

mercury/compiler/mlds_to_c:
	Changed to allow printing of multiple return types for mlds dumps.