Ensure that before attempting to build a parent module of a nested
Estimated hours taken: 3
Branches: main

Ensure that before attempting to build a parent module of a nested
sub-module that we have built all the dependencies of that sub-module.
This is because the call to build the parent module will also build
the nested sub-module so all the dependencies better exist or the build
will fail.

compiler/modules.m:
    Add dependencies between the parent module build targets and the
    sub-module dependencies.