Add two pieces of information that will be needed to an upcoming change
Estimated hours taken: 2
Branches: main

Add two pieces of information that will be needed to an upcoming change
to the representation of procedure bodies in the declarative debugger.
The first piece is the full list of head vars. This is needed now because the
debugger no longer ignores variables of type type_info, and thus the Nth
headvar is no longer guaranteed to be variable number N. The second piece
is a cut/no_cut indicator in `some' goals. This is needed to allow us to
reconstruct goal paths exactly as we traverse goal representations.

browser/program_representation.m:
	Make the two data structure changes described above.

browser/*.m:
	Trivial changes to conform to the new data structures.

compiler/prog_rep.m:
	Create the updated data procedure representation.

compiler/*.m:
	Record and transmit the information needed by the updated procedure
	representation.

runtime/mercury_stack_layout.h:
	Update the documentation of the relevant field in proc_layouts.